GCP_Cloud_Functions/ml/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage==1.42.3
google-cloud-vision==2.6.2
